correcting directory structure mistake
