+++ pi_display/__init__.py
@@ -1,3 +1,4 @@
 '''give module access to display implementations'''
 from .pi_display import PiDisplay
-from .waveshare_eink import EinkDisplay
+#from .waveshare_eink import EinkDisplay
+from .png_display import PngDisplay
