--- pi_display/pi_display/__init__.py
@@ -1,4 +1,0 @@
-'''give module access to display implementations'''
-from .pi_display import PiDisplay
-#from .waveshare_eink import EinkDisplay
-from .png_display import PngDisplay
