+++ pi_display/waveshare_eink.py
@@ -6,10 +6,10 @@ from pi_display import PiDisplay
 
 class EinkDisplay(PiDisplay):
     def __init__(self) -> None:
-        super().__init__()
         logging.basicConfig(level=logging.DEBUG)
         logging.info("Initialising screen...")
         self.epd = epd2in13b_V4.EPD()
+        super().__init__(self, self.epd.height, self.epd.width)
         self.epd.init()
         self.epd.clear()
         time.sleep(1)
